Adicao exercicio sobre area do quadrado e alguns ajustes

+++ eclipse/curso_java_basico/src/com/loiane/cursojava/exercicios/aula13/CelsiusParaFarenheit.java
@@ -6,10 +6,10 @@ public class CelsiusParaFarenheit {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		double farenheit;
-		int celsius;
+		double celsius;
 		
 		System.out.println("Digite a temperatura em celsius");
-		celsius = scan.nextInt();
+		celsius = scan.nextDouble();
 		farenheit = 9*celsius/5 + 32;
 		System.out.println("A temperatura em farenheit é " + farenheit);
 	}
